Update auth handlers and struct_user (#12)
* Add dependencies for web_api crate

* Add upload functionality and update dependencies

* Update file upload path for NFE XML files

* Remove unused XML files from uploads directory

* Add ignore rule for XML files in uploads directory

* Delete web_api/src/public/uploads/nfe/xml

* Update dependencies and add watcher module

* Add migration functionality and update API routes

* Add new modules and routes

* Update dependencies and enable guard middleware

* Update auth handlers and struct_user

This commit updates the `auth_handlers.rs` file in the `web_api/src/services/auth` directory. It adds a new import statement for the `LoginTockenCheckModel` struct in the `struct_user.rs` file. It also modifies the `login_user_post` function to use the `LoginTockenCheckModel` struct instead of the `LoginUserModel` struct for the database query. Additionally, it updates the response model in the `login_user_post` function to include the `firstname` field from the `LoginTockenCheckModel` struct.

* Update route layer middleware to include guard

* Update sleep duration in watcher and fix formatting in web_api

* Refactor imports and add new fields to NfeSelect struct

* Refactor struct field names in nfe/src/modules/json/structs and ipl modules

* Fix typos and update database queries